# openapi.yaml
openapi: 3.1.0
info:
  title: 活動報名系統 API
  version: 0.1.0
  description: |
    18 條 endpoint + /health。一份契約餵 web(Pages)與 iOS(swift-openapi-generator)。
    慣例:金額一律整數「分」(*_cents);時間一律 epoch 毫秒整數;每個 JSON 回應帶 x-server-now header;
    錯誤一律 { error, server_now }。狀態碼:400 輸入、401 身分、403 權限、404 不存在或不屬於你、409 規則擋下。
servers:
  - url: http://127.0.0.1:8788
    description: wrangler dev --local
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      security: []
      summary: 建立成員(單一組織,角色 member)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, nickname]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                password: { type: string, minLength: 8 }
                nickname: { type: string, minLength: 1, maxLength: 50 }
      responses:
        '201': { description: 已建立, content: { application/json: { schema: { $ref: '#/components/schemas/Member' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      security: []
      summary: 登入,回 access(15 分鐘)+ refresh(30 天)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      operationId: refresh
      tags: [auth]
      security: []
      summary: 輪替 refresh;舊的立刻失效;重放 → 撤銷該成員全部 refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200': { description: ok, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: 撤銷當前 refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '204': { description: 已撤銷 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ── 活動(5)─────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    BadRequest:   { description: invalid_input,  content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: unauthorized / refresh_replayed, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    Conflict:     { description: 規則擋下(見 error 代碼), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
    Id: { type: string, minLength: 1 }
    EpochMs: { type: integer, format: int64, minimum: 0, description: epoch 毫秒 }
...
    Role: { type: string, enum: [member, staff] }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum: [invalid_input, unauthorized, refresh_replayed, forbidden, not_found, not_implemented, internal,
                 email_taken, not_on_sale, seat_taken, hold_exists, sold_out, hold_expired, terminal_state,
                 capacity_below_sold, capacity_exceeded, promo_rejected]
        server_now: { $ref: '#/components/schemas/EpochMs' }
    Member:
      type: object
      required: [id, email, nickname, role]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        email: { type: string }
        nickname: { type: string }
        role: { $ref: '#/components/schemas/Role' }
    TokenPair:
      type: object
      required: [access_token, refresh_token, expires_at, member]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_at: { $ref: '#/components/schemas/EpochMs', description: access 到期 }
        member: { $ref: '#/components/schemas/Member' }
...